fixed a float promotion in axoloti math

+++ shared/axoloti-wrapper/axoloti_math.cpp
@@ -60,7 +60,9 @@ void axoloti_math_init(void)
 		uint32_t *q;
 		q = (uint32_t *)pitcht;
 		for (i = 0; i < PITCHTSIZE; i++) {
-			double f = 440.0 * powf(2.0, (i - 69.0 - 64.0) / 12.0);
+			//Fix: powf returns a double, so we use pow() to avoid the a hidden float->double promotion:
+			// double f = 440.0 * powf(2.0, (i - 69.0 - 64.0) / 12.0);
+			double f = 440.0 * pow(2.0, (i - 69.0 - 64.0) / 12.0);
 			double phi = 4.0 * (double)(1 << 30) * f / (SAMPLERATE * 1.0);
 			if (phi > ((unsigned int)1 << 31))
 				phi = 0x7FFFFFFF;
